pyproject.toml
[tool.black]
line-length = 120
target-version = ['py39']

[tool.poetry]
name = "flights-monitor"
version = "0.0.1"
description = ""
authors = ["thisi5patrick"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.0"
selenium = "^4.2.0"
orjson = "^3.7.2"
Flask = "^2.1.2"
Flask-SQLAlchemy = "^2.5.1"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
py_version = 39
profile = "black"
line_length = 120
skip = ["db/tables/__init__.py"]